requirements.txt
# Core Framework
kedro>=0.19.0
kedro-mlflow>=0.13.0
mlflow>=2.10.0

# Deep Learning
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.15.0
onnxruntime-gpu>=1.17.0

# Computer Vision
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
albumentations>=1.4.0
supervision>=0.20.0

# Object Detection
ultralytics>=8.2.0  # YOLOv11/v12
rfdetr>=0.1.0
transformers>=4.38.0

# Tracking
lap>=0.4.0  # Linear Assignment Problem solver
filterpy>=1.4.5  # Kalman filtering
scipy>=1.12.0

# Lane Detection
mmcv>=2.1.0
mmdet>=3.3.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pillow>=10.2.0
pycocotools>=2.0.7

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
streamlit>=1.31.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
loguru>=0.7.2
hydra-core>=1.3.2
omegaconf>=2.3.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Video Processing
decord>=0.6.0
av>=11.0.0

# ReID (for tracking)
faiss-cpu>=1.7.4  # Or faiss-gpu for GPU support
timm>=0.9.12

# Path Planning
bezier>=2023.7.28
shapely>=2.0.2
